% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerCareSentimentR-package.R
\docType{package}
\name{CustomerCareSentimentR-package}
\alias{CustomerCareSentimentR}
\alias{CustomerCareSentimentR-package}
\title{CustomerCareSentimentR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
